fix(timeout): add explicit casts and simplify timeout calculation for nullable TimeoutSeconds
- Add explicit double cast to TimeoutSeconds fallback expressions in all providers to ensure correct TimeSpan.FromSeconds overload
- Extract clamped timeout calculation to separate variable in AIToolCall.ExecuteAsync for clarity
- Update Gemini streaming to use .Value accessor for nullable TimeoutSeconds
- Handle nullable TimeoutSeconds with null-coalescing in AIToolCall timeout resolution

+++ src/SmartHopper.Providers.OpenAI/OpenAIProvider.cs
@@ -1167,7 +1167,7 @@ namespace SmartHopper.Providers.OpenAI
                 }
 
                 // Determine idle timeout from request (fallback to 600s if invalid)
-                var idleTimeout = TimeSpan.FromSeconds(request.TimeoutSeconds > 0 ? request.TimeoutSeconds : 600);
+                var idleTimeout = TimeSpan.FromSeconds((double)(request.TimeoutSeconds > 0 ? request.TimeoutSeconds : 600));
                 await foreach (var data in this.ReadSseDataAsync(
                     response,
                     idleTimeout,
